Create MAC Address Changer
PowerShell script for changing the Media Access Control (MAC) address of a network adapter on a Windows machine.


Description of the primary components of the script:

1. Generate-MAC function
This function generates a new random MAC address. It starts with "02" and then adds 5 more randomly generated hexadecimal values (ranging from 00 to FF), representing the 6-byte structure of a MAC address.

2. IsValid-MAC function
This function checks whether a given MAC address is valid or not. It uses a regular expression to match the standard MAC address format which consists of six groups of two hexadecimal digits, separated by hyphens or colons.

3. Setting up the output directory
The script defines an output directory (C:\MAC Address changes) and creates it if it doesn't already exist.

4. UserPrompt function
This function is used for interacting with the user. It accepts five parameters - a message to be displayed, messages for 'yes' and 'no' responses, and script blocks for 'yes' and 'no' options. It asks the user to enter either 'yes' or 'no', and depending on the input, it runs the corresponding script block and returns true or false.

5. Main function
This is the main function of the script. It includes several steps:
* It first retrieves all network adapters that are currently up on the machine and displays them to the user.
* It then prompts the user to choose a network adapter by entering the interface index (ifIndex) of the adapter.
* Once the adapter is selected, it confirms with the user if they want to proceed with the operation.
* It asks the user if they want to input a specific MAC address or generate a random one. If the user wants to input a MAC address, it checks the validity of the MAC address input by the user.
It then confirms with the user before proceeding to change the MAC address.
* It attempts to change the MAC address by first setting the 'NetworkAddress' property of the network adapter to the new MAC address. It then disables and enables the network adapter to apply the changes.
* After this, it checks if the MAC address change was successful by retrieving the current MAC address of the adapter and comparing it with the new one. If the MAC addresses match, it informs the user of the successful operation, otherwise, it throws an error.
* The script also includes an option to retry the operation or change another MAC address if the user wishes to do so.

6. Calling the Main function
The final line of the script runs the Main function, kicking off the MAC address-changing process.